+++ client/next.config.js
@@ -1,4 +1,8 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+  images: {
+    domains: ['www.batk.nubip.edu.ua'],
+  },
+};
 
 module.exports = nextConfig;