+++ client/src/components/index.ts
@@ -1,0 +1,6 @@
+export * from './article';
+export * from './content';
+export * from './header';
+export * from './pagination';
+export * from './ads';
+export * from './footer';