+++ client/src/shared/constants/routes.ts
@@ -1,0 +1,17 @@
+export const navbarRoutes = {
+  home: {
+    title: 'Головна',
+    path: '/',
+  },
+};
+
+export const otherRoutes = {
+  article: {
+    path: '/article/:id',
+  },
+  category: {
+    path: '/category/:id',
+  },
+};
+
+export const routes = { ...navbarRoutes, ...otherRoutes };