+++ CHANGELOG.md
@@ -2,6 +2,48 @@
 
 Всі важливі зміни в цьому проєкті будуть задокументовані в цьому файлі.
 
+## [0.0.4] - 2023-08-08
+
+### Added
+
+#### Client
+
+- Sticky header
+- Desktop version of the website
+- Pagination
+- Section with ads
+- Footer
+
+### Fixed
+
+#### Client
+
+- Adjusted the width and the height of mobile article image
+- Moved article related code to separate component
+
+## [0.0.3] - 2023-08-02
+
+### Added
+
+#### Client
+
+- Mobile version of the header
+- Mobile version of article card
+
+## [0.0.2] - 2023-08-02
+
+### Added
+
+#### Client
+
+- Mobile version of navbar at the top of the page
+- Navbar drawer to hide links on the mobile
+
+#### Backend
+
+- Main entities of the app
+- Data to have the content to fetch on the client
+
 ## [0.0.1] - 2023-07-31
 
 ### Added